ユースケースで定義
